+++ code/graphics/grstub.cpp
@@ -545,7 +545,7 @@ void gr_stub_shadow_map_end()
 {
 }
 
-void gr_stub_render_model(model_material* material_info, indexed_vertex_source *vert_source, vertex_buffer* bufferp, int texi)
+void gr_stub_render_model(model_material* material_info, indexed_vertex_source *vert_source, vertex_buffer* bufferp, size_t texi)
 {
 
 }
